requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
Deprecated==1.2.15
distro==1.9.0
Django==5.1.3
django-cors-headers==4.6.0
django-modeltranslation==0.19.11
django-unfold==0.42.0
idna==3.10
netaddr==1.3.0
pillow==11.0.0
pycparser==2.22
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
sqlparse==0.5.2
tzdata==2024.2
urllib3==2.3.0
wrapt==1.17.2
yookassa==3.4.3
